django_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.0
channels==4.0.0
daphne==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
channels-redis==4.1.0

# Audio Processing Dependencies
celery==5.3.4
redis==5.0.1
librosa==0.10.1
webrtcvad==2.0.10
openai-whisper==20231117
pydub==0.25.1
noisereduce==3.0.0
pyannote.audio==3.1.1
soundfile==0.12.1
torch==2.1.2
torchaudio==2.1.2

# AI Assistant Dependencies
httpx==0.25.2
uuid==1.30
